+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 # 2.2.3
 * bugfix R-Precision
+* make it possible to save PR-Curve with KWSError
 
 # 2.2.2
 * bugfix KeywordExtractor: Also put Character Class M,Cs,Co to L